震央分布図の3D表示をやめる
描画が遅くなるから。
古いPCだとめっちゃ遅くて固まる。
